refactored chat endpoints to chat controller
